services/workflow-registry/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyyaml==6.0.1
jsonschema==4.20.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2